// src/app/@api/model/itemCategoryDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemSubcategoryDto } from './itemSubcategoryDto';


export interface ItemCategoryDto { 
    categoryId?: number;
    categoryName?: string;
    itemSubcategories?: Array<ItemSubcategoryDto>;
}